Extract the Part Numbers

numextract takes the part numbers from cfr_part output.

numextract(string, return = c("min", "max")[1])

Arguments

string
ideally cfr_part()$parts

return
min or max, default is "min"

Value

numeric value from 1 to `Inf``

Details

numextract

Each CFR chapter has multiple parts that often span volumes. To facilitate targeting a specific part, it's necessary to evaluate which parts are in each volume (e.g., Parts 18 to 199). The CFR sometimes uses terms like "END" or "end" to denote the maximum part in each chapter. numextract simply returns the max as `Inf` in these situations.